scripts/installer/requirements.txt
python-hcl2>=8,<9
